move NetModal to separate file, add to Game.tsx
