+++ src/renderer/Settings.tsx
@@ -1,13 +1,14 @@
-import React, { useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 
 import { Link } from 'react-router-dom';
 import FrenzyNETHeader from './FrenzyNETHeader';
-import { userSettings, modalState } from './interfaces';
+import { modalState } from './interfaces';
 import { useSettingsStore } from '../hooks/useSettingsStore';
+import NetModal from './NetModal';
 
 export default function Settings() {
   const { settings, updateSettings } = useSettingsStore();
-  const [modalState, setModalState] = useState<modalState>({
+  const [localModalState, setLocalModalState] = useState<modalState>({
     title: 'TEST',
     desc: 'Testing modal state.',
     input: 'textarea',
@@ -17,7 +18,7 @@ export default function Settings() {
   });
 
   function closeModal() {
-    setModalState((prev) => ({
+    setLocalModalState((prev) => ({
       ...prev,
       visible: false,
     }));
@@ -67,7 +68,7 @@ export default function Settings() {
               <b>player_theme:</b> {settings.player_theme}{' '}
               <a
                 onClick={() =>
-                  setModalState({
+                  setLocalModalState({
                     title: 'Change Theme',
                     desc: 'Change the FWF player theme.',
                     input: 'dropdown',
@@ -115,7 +116,7 @@ export default function Settings() {
               <b>skip_timer:</b> {settings.skip_timer}{' '}
               <a
                 onClick={() =>
-                  setModalState({
+                  setLocalModalState({
                     title: 'Change Skip Timer',
                     desc: 'Set the duration that you need to wait until the skip button shows up.',
                     input: 'number',
@@ -135,7 +136,7 @@ export default function Settings() {
               <b>username:</b> {settings.username ? settings.username : 'NONE'}{' '}
               <a
                 onClick={() =>
-                  setModalState({
+                  setLocalModalState({
                     title: 'Change Username',
                     desc: 'Change displayed username when using Classic theme.',
                     input: 'text',
@@ -152,7 +153,7 @@ export default function Settings() {
               <b>class:</b> {settings.class ? settings.class : 'NONE'}{' '}
               <a
                 onClick={() =>
-                  setModalState({
+                  setLocalModalState({
                     title: 'Change Class',
                     desc: 'Change your user class when using Classic theme.',
                     input: 'dropdown',
@@ -169,7 +170,7 @@ export default function Settings() {
               <b>location:</b> {settings.location ? settings.location : 'NONE'}{' '}
               <a
                 onClick={() =>
-                  setModalState({
+                  setLocalModalState({
                     title: 'Change Location',
                     desc: 'Change your user location for certain location-based game effects.',
                     input: 'dropdown',
@@ -200,7 +201,7 @@ export default function Settings() {
               <b>volume_master:</b> {settings.volume_master}{' '}
               <a
                 onClick={() =>
-                  setModalState({
+                  setLocalModalState({
                     title: 'Change Master Volume',
                     input: 'number',
                     button: '✓ CHANGE MASTER VOLUME',
@@ -216,7 +217,7 @@ export default function Settings() {
               <b>volume_video:</b> {settings.volume_video}{' '}
               <a
                 onClick={() =>
-                  setModalState({
+                  setLocalModalState({
                     title: 'Change Video Volume',
                     input: 'number',
                     button: '✓ CHANGE VIDEO VOLUME',
@@ -232,7 +233,7 @@ export default function Settings() {
               <b>volume_music:</b> {settings.volume_music}{' '}
               <a
                 onClick={() =>
-                  setModalState({
+                  setLocalModalState({
                     title: 'Change Music Volume',
                     input: 'number',
                     button: '✓ CHANGE MUSIC VOLUME',
@@ -247,266 +248,10 @@ export default function Settings() {
           </div>
         </div>
       </div>
-      <NetModal modalState={modalState} setter={setModalState}></NetModal>
+      <NetModal
+        modalState={localModalState}
+        setter={setLocalModalState}
+      ></NetModal>
     </div>
   );
 }
-
-interface ModalProps {
-  modalState: modalState;
-  setter: React.Dispatch<React.SetStateAction<modalState>>;
-}
-
-function NetModal(props: ModalProps) {
-  const { settings, updateSettings } = useSettingsStore();
-  const modalValue = useRef<HTMLInputElement>(null);
-  const selectValue = useRef<HTMLSelectElement>(null);
-  useEffect(() => {
-    if (modalValue.current) {
-      modalValue.current.value =
-        settings[props.modalState.value as keyof userSettings].toString();
-    }
-    if (selectValue.current) {
-      selectValue.current.value =
-        settings[props.modalState.value as keyof userSettings].toString();
-    }
-  }, [props.modalState.value]);
-
-  function closeModal() {
-    props.setter((prev) => ({
-      ...prev,
-      visible: false,
-    }));
-  }
-
-  function changeSetting(s: string, v: string | boolean) {
-    updateSettings({
-      ...settings,
-      [s]: v,
-    });
-    closeModal();
-  }
-
-  switch (props.modalState.input) {
-    case 'text':
-      return (
-        <div
-          className="NETmodal"
-          style={
-            props.modalState.visible ? { opacity: '100%', display: 'flex' } : {}
-          }
-        >
-          <div className="NETmodalcontainer">
-            <div className="NETmodaltitle">{props.modalState.title}</div>
-            <div className="NETmodaldesc">{props.modalState.desc}</div>
-            <div className="NETmodalinput">
-              <input
-                type="text"
-                id="NETmodalvalue"
-                ref={modalValue}
-                defaultValue={
-                  settings[
-                    props.modalState.value as keyof userSettings
-                  ] as string
-                }
-              ></input>
-            </div>
-            <div className="NETmodalbuttons">
-              <a onClick={closeModal}>
-                <div>× CANCEL</div>
-              </a>
-              <a
-                className="purple"
-                onClick={() => {
-                  const value = document.getElementById(
-                    'NETmodalvalue',
-                  ) as HTMLInputElement;
-                  changeSetting(props.modalState.value, value.value);
-                }}
-              >
-                <div>{props.modalState.button}</div>
-              </a>
-            </div>
-          </div>
-        </div>
-      );
-    case 'number':
-      return (
-        <div
-          className="NETmodal"
-          style={
-            props.modalState.visible ? { opacity: '100%', display: 'flex' } : {}
-          }
-        >
-          <div className="NETmodalcontainer">
-            <div className="NETmodaltitle">{props.modalState.title}</div>
-            <div className="NETmodaldesc">{props.modalState.desc}</div>
-            <div className="NETmodalinput">
-              <input
-                type="number"
-                id="NETmodalvalue"
-                ref={modalValue}
-                defaultValue={
-                  settings[
-                    props.modalState.value as keyof userSettings
-                  ] as string
-                }
-              ></input>
-            </div>
-            <div className="NETmodalbuttons">
-              <a onClick={closeModal}>
-                <div>× CANCEL</div>
-              </a>
-              <a
-                className="purple"
-                onClick={() => {
-                  const value = document.getElementById(
-                    'NETmodalvalue',
-                  ) as HTMLInputElement;
-                  changeSetting(props.modalState.value, value.value);
-                }}
-              >
-                <div>{props.modalState.button}</div>
-              </a>
-            </div>
-          </div>
-        </div>
-      );
-    case 'dropdown':
-      let dropdown;
-      switch (props.modalState.value) {
-        case 'class':
-          dropdown = classDropdown;
-          break;
-        case 'location':
-          dropdown = locationDropdown;
-          break;
-        case 'player_theme':
-          dropdown = themeDropdown;
-          break;
-        default:
-          dropdown = defaultDropdown;
-          break;
-      }
-      return (
-        <div
-          className="NETmodal"
-          style={
-            props.modalState.visible ? { opacity: '100%', display: 'flex' } : {}
-          }
-        >
-          <div className="NETmodalcontainer">
-            <div className="NETmodaltitle">{props.modalState.title}</div>
-            <div className="NETmodaldesc">{props.modalState.desc}</div>
-            <div className="NETmodalinput">
-              <select
-                id="NETmodalvalue"
-                ref={selectValue}
-                defaultValue={
-                  settings[
-                    props.modalState.value as keyof userSettings
-                  ] as string
-                }
-              >
-                {dropdown.map((option: dropdownOption) => (
-                  <option key={option.value} value={option.value}>
-                    {option.label}
-                  </option>
-                ))}
-              </select>
-            </div>
-            <div className="NETmodalbuttons">
-              <a onClick={closeModal}>
-                <div>× CANCEL</div>
-              </a>
-              <a
-                className="purple"
-                onClick={() => {
-                  const value = document.getElementById(
-                    'NETmodalvalue',
-                  ) as HTMLInputElement;
-                  changeSetting(props.modalState.value, value.value);
-                }}
-              >
-                <div>{props.modalState.button}</div>
-              </a>
-            </div>
-          </div>
-        </div>
-      );
-    default:
-      return null;
-  }
-}
-
-interface dropdownOption {
-  value: string;
-  label: string;
-}
-
-const defaultDropdown: dropdownOption[] = [{ value: 'NONE', label: 'NONE' }];
-
-const themeDropdown: dropdownOption[] = [
-  { value: 'classic', label: 'classic' },
-];
-
-const classDropdown: dropdownOption[] = [
-  { value: 'NONE', label: 'NONE' },
-  { value: '#', label: 'Regulator' },
-  { value: '$', label: 'Banker' },
-  { value: '*', label: 'Senator' },
-];
-
-const locationDropdown: dropdownOption[] = [
-  { value: 'NONE', label: 'NONE' },
-  { value: 'AL', label: 'AL' },
-  { value: 'AK', label: 'AK' },
-  { value: 'AZ', label: 'AZ' },
-  { value: 'AR', label: 'AR' },
-  { value: 'CA', label: 'CA' },
-  { value: 'CO', label: 'CO' },
-  { value: 'CT', label: 'CT' },
-  { value: 'DE', label: 'DE' },
-  { value: 'FL', label: 'FL' },
-  { value: 'GA', label: 'GA' },
-  { value: 'HI', label: 'HI' },
-  { value: 'ID', label: 'ID' },
-  { value: 'IL', label: 'IL' },
-  { value: 'IN', label: 'IN' },
-  { value: 'IA', label: 'IA' },
-  { value: 'KS', label: 'KS' },
-  { value: 'KY', label: 'KY' },
-  { value: 'LA', label: 'LA' },
-  { value: 'ME', label: 'ME' },
-  { value: 'MD', label: 'MD' },
-  { value: 'MA', label: 'MA' },
-  { value: 'MI', label: 'MI' },
-  { value: 'MN', label: 'MN' },
-  { value: 'MS', label: 'MS' },
-  { value: 'MO', label: 'MO' },
-  { value: 'MT', label: 'MT' },
-  { value: 'NE', label: 'NE' },
-  { value: 'NV', label: 'NV' },
-  { value: 'NH', label: 'NH' },
-  { value: 'NJ', label: 'NJ' },
-  { value: 'NM', label: 'NM' },
-  { value: 'NY', label: 'NY' },
-  { value: 'NC', label: 'NC' },
-  { value: 'ND', label: 'ND' },
-  { value: 'OH', label: 'OH' },
-  { value: 'OK', label: 'OK' },
-  { value: 'OR', label: 'OR' },
-  { value: 'PA', label: 'PA' },
-  { value: 'RI', label: 'RI' },
-  { value: 'SC', label: 'SC' },
-  { value: 'SD', label: 'SD' },
-  { value: 'TN', label: 'TN' },
-  { value: 'TX', label: 'TX' },
-  { value: 'UT', label: 'UT' },
-  { value: 'VT', label: 'VT' },
-  { value: 'VA', label: 'VA' },
-  { value: 'WA', label: 'WA' },
-  { value: 'WV', label: 'WV' },
-  { value: 'WI', label: 'WI' },
-  { value: 'WY', label: 'WY' },
-];
